creating packages and moving suitable classes inside
