+++ src/main/java/xoGame/xoGameExceptions/MyExceptionHandler.java
@@ -1,4 +1,4 @@
-package xoGame;
+package xoGame.xoGameExceptions;
 
 public class MyExceptionHandler {
